+++ src/App.tsx
@@ -1,11 +1,12 @@
-import { Loader } from './components/Loader';
-import { Link, Routes, Route, Navigate, useLocation } from 'react-router-dom';
+import {
+  NavLink,
+  Routes,
+  Route,
+  Navigate,
+  useLocation,
+} from 'react-router-dom';
 import './App.scss';
-import { getPeople } from './api';
-import { useEffect, useState } from 'react';
-import { Person } from './types';
-
-const nomesColunas = ['Name', 'Sex', 'Born', 'Died', 'Mother', 'Father'];
+import { PeoplePage } from './components/PeopleTable';
 
 const HomePage = () => (
   <div className="container">
@@ -18,135 +19,6 @@ const PageNotFound = () => (
     <h1 className="title">Page not found</h1>
   </div>
 );
-
-type PersonLinkProps = {
-  person: Person;
-  onSelect: (slug: string) => void;
-};
-
-const PersonLink = ({ person, onSelect }: PersonLinkProps) => {
-  if (!person.slug) {
-    return null;
-  }
-
-  return (
-    <Link
-      to={`/people/${person.slug}`}
-      onClick={() => onSelect(person.slug)}
-      className={person.sex === 'f' ? 'has-text-danger' : ''}
-    >
-      {person.name}
-    </Link>
-  );
-};
-
-const PeoplePage = () => {
-  const [people, setPeople] = useState<Person[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState(false);
-  const [seletedSlug, setSeletedSlug] = useState<string | null>(null);
-
-  useEffect(() => {
-    getPeople()
-      .then(data => {
-        setPeople(data);
-        setError(false);
-      })
-      .catch(() => setError(true))
-      .finally(() => setIsLoading(false));
-  }, []);
-
-  return (
-    <div className="container">
-      <h1 className="title">People Page</h1>
-
-      <div className="block">
-        <div className="box table-container">
-          {isLoading && <Loader />}
-
-          {error && (
-            <p data-cy="peopleLoadingError" className="has-text-danger">
-              Something went wrong
-            </p>
-          )}
-
-          {!isLoading && !error && people.length === 0 && (
-            <p data-cy="noPeopleMessage">There are no people on the server</p>
-          )}
-
-          {people.length > 0 && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  {nomesColunas.map(coluna => (
-                    <th key={coluna}>{coluna}</th>
-                  ))}
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => {
-                  const mother = people.find(p => p.name === person.motherName);
-                  const father = people.find(p => p.name === person.fatherName);
-
-                  return (
-                    <tr
-                      data-cy="person"
-                      key={person.slug}
-                      className={
-                        seletedSlug === person.slug
-                          ? 'has-background-warning'
-                          : ''
-                      }
-                    >
-                      <td>
-                        <PersonLink person={person} onSelect={setSeletedSlug} />
-                      </td>
-                      <td>{person.sex}</td>
-                      <td>{person.born}</td>
-                      <td>{person.died}</td>
-                      <td>
-                        {person.motherName ? (
-                          mother ? (
-                            <PersonLink
-                              person={mother}
-                              onSelect={setSeletedSlug}
-                            />
-                          ) : (
-                            person.motherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                      <td>
-                        {person.fatherName ? (
-                          father ? (
-                            <PersonLink
-                              person={father}
-                              onSelect={setSeletedSlug}
-                            />
-                          ) : (
-                            person.fatherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
 
 export const App = () => {
   const { pathname } = useLocation();
@@ -161,21 +33,21 @@ export const App = () => {
       >
         <div className="container">
           <div className="navbar-brand">
-            <Link
+            <NavLink
               className={`navbar-item
               ${pathname === '/' && 'has-background-grey-lighter'}`}
               to="/"
             >
               Home
-            </Link>
+            </NavLink>
 
-            <Link
+            <NavLink
               className={`navbar-item
               ${pathname === '/people' && 'has-background-grey-lighter'}`}
               to="/people"
             >
               People
-            </Link>
+            </NavLink>
           </div>
         </div>
       </nav>
